+++ crates/gdcef/src/cef_texture/backend.rs
@@ -1,8 +1,11 @@
 use adblock::lists::{FilterSet, ParseOptions};
 use cef::{BrowserSettings, ImplBrowser, ImplBrowserHost, RequestContextSettings, WindowInfo};
 use cef_app::PhysicalSize;
+use godot::classes::Image;
+use godot::classes::image::Format as ImageFormat;
 use godot::classes::{AudioServer, DisplayServer, Engine, ImageTexture, Texture2Drd};
 use godot::prelude::*;
+use software_render::{DestBuffer, PopupBuffer, composite_popup};
 use std::collections::HashMap;
 use std::fs;
 use std::rc::Rc;
@@ -14,6 +17,7 @@ use crate::accelerated_osr::{
 };
 use crate::browser::{App, BrowserState, PopupPolicyFlag, PopupStateQueue, RenderMode};
 use crate::error::CefError;
+use crate::utils::get_display_scale_factor;
 use crate::{godot_protocol, render, webrender};
 
 /// Shared browser creation inputs used by both `CefTexture` and `CefTexture2D`.
@@ -175,6 +179,148 @@ pub(crate) fn apply_popup_policy(app: &App, policy: i32) {
     if let Some(state) = app.state.as_ref() {
         state.popup_policy.store(policy, Ordering::Relaxed);
     }
+}
+
+/// Updates the primary browser texture for both software and accelerated modes.
+///
+/// Returns a replacement accelerated texture when a resize occurs. `CefTexture`
+/// uses this to update its `TextureRect` texture binding immediately.
+pub(crate) fn update_primary_texture(
+    state: &mut BrowserState,
+    log_prefix: &str,
+) -> Option<Gd<Texture2Drd>> {
+    if let RenderMode::Software {
+        frame_buffer,
+        texture,
+    } = &mut state.render_mode
+    {
+        let Ok(mut fb) = frame_buffer.lock() else {
+            return None;
+        };
+
+        let popup_metadata = state.popup_state.lock().ok().and_then(|popup| {
+            if popup.visible && !popup.buffer.is_empty() {
+                Some((
+                    popup.width,
+                    popup.height,
+                    popup.rect.x,
+                    popup.rect.y,
+                    popup.dirty,
+                ))
+            } else {
+                None
+            }
+        });
+
+        let popup_dirty = popup_metadata
+            .as_ref()
+            .is_some_and(|(_, _, _, _, dirty)| *dirty);
+
+        if !fb.dirty && !popup_dirty {
+            return None;
+        }
+
+        if fb.data.is_empty() {
+            return None;
+        }
+
+        let width = fb.width as i32;
+        let height = fb.height as i32;
+        let display_scale = get_display_scale_factor();
+
+        let final_data =
+            if let Some((popup_width, popup_height, popup_x, popup_y, _)) = popup_metadata {
+                let popup_buffer = state
+                    .popup_state
+                    .lock()
+                    .ok()
+                    .map(|popup| popup.buffer.clone());
+
+                if let Some(popup_buffer) = popup_buffer {
+                    let mut composited = fb.data.clone();
+                    let scaled_x = (popup_x as f32 * display_scale) as i32;
+                    let scaled_y = (popup_y as f32 * display_scale) as i32;
+                    composite_popup(
+                        &mut DestBuffer {
+                            data: &mut composited,
+                            width: fb.width,
+                            height: fb.height,
+                        },
+                        &PopupBuffer {
+                            data: &popup_buffer,
+                            width: popup_width,
+                            height: popup_height,
+                            x: scaled_x,
+                            y: scaled_y,
+                        },
+                    );
+                    if let Ok(mut popup) = state.popup_state.lock() {
+                        popup.mark_clean();
+                    }
+                    composited
+                } else {
+                    fb.data.clone()
+                }
+            } else {
+                fb.data.clone()
+            };
+
+        let byte_array = PackedByteArray::from(final_data.as_slice());
+        let image: Option<Gd<Image>> =
+            Image::create_from_data(width, height, false, ImageFormat::RGBA8, &byte_array);
+        if let Some(image) = image {
+            texture.set_image(&image);
+        }
+
+        fb.mark_clean();
+        return None;
+    }
+
+    #[cfg(any(target_os = "macos", target_os = "windows", target_os = "linux"))]
+    if let RenderMode::Accelerated {
+        render_state,
+        texture_2d_rd,
+    } = &mut state.render_mode
+    {
+        let Ok(mut accel_state) = render_state.lock() else {
+            return None;
+        };
+
+        let texture_to_set = if let Some((new_w, new_h)) = accel_state.needs_resize.take()
+            && new_w > 0
+            && new_h > 0
+        {
+            render::free_rd_texture(accel_state.dst_rd_rid);
+
+            let (new_rd_rid, new_texture_2d_rd) =
+                match render::create_rd_texture(new_w as i32, new_h as i32) {
+                    Ok(result) => result,
+                    Err(e) => {
+                        godot::global::godot_error!("[{}] {}", log_prefix, e);
+                        return None;
+                    }
+                };
+
+            accel_state.dst_rd_rid = new_rd_rid;
+            accel_state.dst_width = new_w;
+            accel_state.dst_height = new_h;
+
+            *texture_2d_rd = new_texture_2d_rd.clone();
+            Some(new_texture_2d_rd)
+        } else {
+            None
+        };
+
+        if accel_state.has_pending_copy
+            && let Err(e) = accel_state.process_pending_copy()
+        {
+            godot::global::godot_error!("[{}] Failed to process pending copy: {}", log_prefix, e);
+        }
+
+        return texture_to_set;
+    }
+
+    None
 }
 
 pub(crate) fn try_create_browser(
